Refs #2330: Migrates ElggPriorityList into its own PHPUnit test case and runs it automatically on Travis.
