+++ .travis.yml
@@ -1,9 +1,6 @@
 language: php
 phps:
- - 5.2
  - 5.3
-
-script: phpunit --help
 
 branches:
   only:
