+++ engine/tests/phpunit/bootstrap.php
@@ -1,0 +1,7 @@
+<?php
+
+$engine = dirname(dirname(__DIR__));
+
+// Set up class auto-loading
+require_once "$engine/lib/elgglib.php";
+require_once "$engine/lib/autoloader.php";
