+++ phpunit.xml
@@ -1,0 +1,7 @@
+<phpunit bootstrap="./tests/phpunit/bootstrap.php">
+    <testsuites>
+		<testsuite name="All Tests">
+    			<directory>./engine/tests/phpunit/</directory>
+		</testsuite>
+	</testsuites>
+</phpunit>
